requirements.txt
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
